lots of changes to content fetching
