+++ apps/api/src/modules/content/db/schemas/poetry-content.schema.ts
@@ -6,13 +6,13 @@ import {
     WorkStatus,
     Genres,
     ContentRating,
-    SectionInfo,
     ContentKind,
     PubStatus,
     PoetryForm,
     TagsModel,
 } from '$shared/models/content';
 import { Pseudonym } from '$shared/models/accounts';
+import { Section } from '$shared/models/sections';
 
 @Schema()
 export class PoetryContentDocument extends Document implements PoetryContent {
@@ -59,7 +59,7 @@ export class PoetryContentDocument extends Document implements PoetryContent {
         ref: 'Sections',
         default: null,
     })
-    sections?: string[] | SectionInfo[];
+    sections?: string[] | Section[];
 
     @Prop({ type: [{ type: String, ref: 'Tags' }] })
     tags?: string[] | TagsModel[];
