+++ apps/api/src/modules/content/services/content.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, UnauthorizedException } from '@nestjs/common';
 import { PaginateResult } from 'mongoose';
 import {
     BlogsStore,
