+++ apps/api/src/modules/content/services/sections.service.ts
@@ -18,6 +18,10 @@ export class SectionsService {
         private readonly events: EventEmitter2,
     ) {}
 
+    async fetchSections(contentId: string, published = false, userId?: string) {
+        return await this.content.fetchSections(contentId, published);
+    }
+
     async fetchOneById(sectionId: string, published?: boolean): Promise<Section> {
         return await this.content.fetchSectionById(sectionId, published);
     }
