+++ apps/client/src/lib/components/ui/misc/index.ts
@@ -4,3 +4,4 @@ export { default as Jumbotron } from './Jumbotron.svelte';
 export { default as Paginator } from './Paginator.svelte';
 export { default as Time } from './Time.svelte';
 export { default as Toast } from './Toast.svelte';
+export { default as NotifyBanner } from './NotifyBanner.svelte';
